# services/api-gateway/internal/docs/openapi/openapi-organization.yaml
openapi: 3.0.3
info:
  title: CivicOS Organization Service
  version: 1.0.0
  description: |
    Organizations (registry + membership), announcements, projects, issue
    assignments ("receive reports"), and progress updates ("respond publicly",
    "update progress").

    All routes are proxied by the API gateway at `/api/v1/*`. The service
    itself mounts them under `/v1/*`.

    ## Authorization model

    Roles allowed to **create** a top-level org from scratch: `GOVERNMENT_ADMIN`,
    `PLATFORM_ADMIN`, `NGO`. Once created, the org's internal role
    (`OWNER` / `ADMIN` / `STAFF`) governs edits — not the platform JWT role.

    **Content-authorship writes are strictly org-only.** A `PLATFORM_ADMIN`
    who is not a member of the org **cannot** create, edit, publish, or
    delete announcements / projects / assignments / consultations on the
    org's behalf. Attribution belongs to the organization itself.

    Platform admins retain two narrow powers:

    - **Emergency close/archive** — `PLATFORM_ADMIN` can close a published
      consultation or archive an announcement without joining the org first.
      Written to the audit log with the actor's identity.
    - **Oversight reads** — `PLATFORM_ADMIN` can view responses, analytics,
      and org-only content (draft announcements, assignment inboxes) for
      any org.

    Anything beyond these two levers requires the platform admin to be
    added as a member of the target org first.
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssignmentStatus:
      type: string
      enum: [RECEIVED, IN_PROGRESS, COMPLETED, REJECTED]
...
paths:
...
  /organizations/{id}/assignments:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Assignments]
      summary: List an organization's issue assignments (members only)
      description: Members-only to prevent a curious user from enumerating another org's inbox.
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/AssignmentStatus' }
      responses:
        '200': { description: Assignments. }
    post:
      tags: [Assignments]
      summary: Claim an issue for an organization (org admin or PLATFORM_ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issueId]
              properties:
                issueId: { type: string, format: uuid }
                note: { type: string, nullable: true }
            examples:
              default:
                {
                  value:
                    {
                      issueId: 7f8e9d0c-1234-5678-9abc-def012345678,
                      note: We've dispatched a team to Junction 4.,
                    },
                }
      responses:
        '201': { description: Assignment created. }
  /issues/{issueId}/assignments:
    parameters:
      - in: path
        name: issueId
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Assignments]
      security: []
      summary: Which orgs have claimed this issue
      description: Public read — anyone viewing an issue should see who's responsible.
      responses:
        '200': { description: Assignments. }
  /assignments/{assignmentId}:
    parameters:
      - in: path
        name: assignmentId
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Assignments]
      summary: Update assignment status (org admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/AssignmentStatus' }
                note: { type: string, nullable: true }
            examples:
              default: { value: { status: IN_PROGRESS, note: Team on site since 8am. } }
      responses:
        '200': { description: Updated assignment. }
    delete:
      tags: [Assignments]
      summary: Remove an assignment (org admin)
      responses:
        '200': { description: Removed. }

  # ─────────────────── PROGRESS UPDATES ───────────────────
...